requirements.txt
# Core Django (4.2 LTS - compatible with Python 3.14)
Django==4.2.10
django-leaflet==0.30.1

# Database (PostgreSQL only needed for production deployment)
# psycopg2-binary is removed for local dev - we'll use SQLite
dj-database-url==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.2

# Production Server
whitenoise==6.6.0
gunicorn==21.2.0

# Utilities
requests==2.31.0
geopy==2.4.1
